Fixed small bug with uploads (fixed earlier but somehow got erased)

+++ app/controllers/uploads_controller.rb
@@ -75,7 +75,9 @@ class UploadsController < ApplicationController
   # DELETE /uploads/1.json
   def destroy
     @upload = Upload.find(params[:id])
-	File.delete(Rails.root.join('..', 'uploads', @upload.file_name))
+	if File.exists?(Rails.root.join('..', 'uploads', @upload.file_name))
+		File.delete(Rails.root.join('..', 'uploads', @upload.file_name))	
+	end
     @upload.destroy
 
     respond_to do |format|
